+++ src/navigators/AppNavigator.tsx
@@ -4,18 +4,7 @@ import AuthStackNavigator from './AuthStack';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import Tabs from './Tabs';
 import {RootStackParamList} from './AppParamList';
-export const navigationRef = createNavigationContainerRef<RootStackParamList>();
-
-export function navigate<T extends keyof RootStackParamList>(
-    name: T,
-    params?: RootStackParamList[T],
-) {
-    if (navigationRef.isReady()) {
-        navigationRef.navigate(name, params);
-    } else {
-        console.error('navigationRef not ready');
-    }
-}
+import {navigationRef} from './navigationUtilities';
 
 export default function AppNavigator() {
     const Stack = createNativeStackNavigator();
