+++ src/navigators/navigationUtilities.ts
@@ -1,0 +1,63 @@
+import {useState, useEffect, useRef} from 'react';
+import {BackHandler, Linking, Platform} from 'react-native';
+import {
+    NavigationState,
+    PartialState,
+    createNavigationContainerRef,
+} from '@react-navigation/native';
+import {RootStackParamList} from './AppParamList';
+
+/**
+ * Reference to the root App Navigator.
+ *
+ * If needed, you can use this to access the navigation object outside of a
+ * `NavigationContainer` context. However, it's recommended to use the `useNavigation` hook whenever possible.
+ * @see [Navigating Without Navigation Prop]{@link https://reactnavigation.org/docs/navigating-without-navigation-prop/}
+ *
+ * The types on this reference will only let you reference top level navigators. If you have
+ * nested navigators, you'll need to use the `useNavigation` with the stack navigator's ParamList type.
+ */
+export const navigationRef = createNavigationContainerRef<RootStackParamList>();
+
+/**
+ * use this to navigate without the navigation
+ * prop. If you have access to the navigation prop, do not use this.
+ * @see {@link https://reactnavigation.org/docs/navigating-without-navigation-prop/}
+ * @param {unknown} name - The name of the route to navigate to.
+ * @param {unknown} params - The params to pass to the route.
+ */
+
+export function navigate<T extends keyof RootStackParamList>(
+    name: T,
+    params?: RootStackParamList[T],
+) {
+    if (navigationRef.isReady()) {
+        navigationRef.navigate(name, params);
+    } else {
+        console.error('navigationRef not ready');
+    }
+}
+/**
+ * This function is used to go back in a navigation stack, if it's possible to go back.
+ * If the navigation stack can't go back, nothing happens.
+ * The navigationRef variable is a React ref that references a navigation object.
+ * The navigationRef variable is set in the App component.
+ */
+export function goBack() {
+    if (navigationRef.isReady() && navigationRef.canGoBack()) {
+        navigationRef.goBack();
+    }
+}
+
+/**
+ * resetRoot will reset the root navigation state to the given params.
+ * @param {Parameters<typeof navigationRef.resetRoot>[0]} state - The state to reset the root to.
+ * @returns {void}
+ */
+export function resetRoot(
+    state: Parameters<typeof navigationRef.resetRoot>[0] = {index: 0, routes: []},
+) {
+    if (navigationRef.isReady()) {
+        navigationRef.resetRoot(state);
+    }
+}
